requirements.txt
apiview==1.3.25
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.2
django-allauth==0.44.0
django-cors-headers==3.2.1
django-import-export==2.0.1
django-smtp-ssl==1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==19.7.1
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.3
password-strength==0.0.3.post2
Pillow==7.0.0
protobuf==3.11.2
psycopg2==2.8.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
typing-extensions==3.10.0.0
urllib3==1.25.8
views==0.3
xlrd==1.2.0
xlwt==1.3.0
